experiments/requirements.txt
openai>=1.0.0
psycopg2-binary>=2.9.0
httpx>=0.28.0
pandas>=2.2.0
matplotlib>=3.9.0
scipy>=1.14.0
pytest>=8.3.0
rich>=13.9.0
